Improve and refactor tiling algorithm
