+++ crates/term-wm-core/src/command_menu/command_registry.rs
