+++ crates/term-wm-core/src/command_menu/mod.rs
@@ -1,11 +1,11 @@
 pub mod arena;
+pub mod command_registry;
 pub mod context;
 pub mod event_bus;
 pub mod matcher;
-pub mod registry;
 
 pub use arena::{CommandAction, CommandName, CommandNode, CommandNodeId, ComponentId};
+pub use command_registry::CommandRegistry;
 pub use context::ContextMask;
 pub use event_bus::{CommandMenuEvent, CommandMenuEventBus};
 pub use matcher::{FuzzyMatch, MruRanker};
-pub use registry::CommandRegistry;
