+++ crates/term-wm-core/src/components.rs
