+++ crates/term-wm-core/src/layout/mod.rs
@@ -1,21 +1,12 @@
-// TODO: Many aspects of this should likely be refactored out of the core, especially which involve rendering
-
 pub mod floating;
 pub mod tiling;
 
+pub use term_wm_layout_engine::Direction;
 pub use tiling::TilingLayout;
 pub use tiling::{InsertPosition, LayoutNode, LayoutPlan, SplitHandle};
 
 use crate::Rect;
 use std::collections::BTreeMap;
-
-/// Layout direction — pure data, no rendering dependency.
-#[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
-pub enum Direction {
-    #[default]
-    Horizontal,
-    Vertical,
-}
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum RectSpec {
@@ -137,7 +128,6 @@ mod tests {
                 height: 4
             }
         );
-
         let pct = RectSpec::Percent {
             x: 50,
             y: 50,
@@ -145,7 +135,6 @@ mod tests {
             height: 50,
         };
         let r = pct.resolve(area);
-        // 50% of width=200 is 100; x offset 50% -> 100 + area.x
         assert_eq!(r.width, 100);
         assert_eq!(r.height, 50);
     }
@@ -169,9 +158,7 @@ mod tests {
         map.set(2u8, b);
         assert_eq!(map.get(1u8), Some(a));
         assert_eq!(map.ids(), vec![1u8, 2u8]);
-        // hit inside first
         assert_eq!(map.hit_test(2, 2, &[1u8, 2u8]), Some(1u8));
-        // miss both
         assert_eq!(map.hit_test(100, 100, &[1u8, 2u8]), None);
     }
